Discard subexpressions right when they have been returned
Instead of storing a list of all the expressions and then modifying them in-place,
process them once they have been returned from the parser and before putting
them on the expression list. This way we can get rid of all the nodes we don't need.

This also allows us to return nil and skip saving the expression entirelly